pyproject.toml
[tool.poetry]
name = "apache-kafka"
version = "0.1.0"
description = "Your description goes here"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"